requirements.txt
django-cors-headers
djangorestframework-simplejwt 
reportlab
Pillow 